+++ CHANGELOG.md
@@ -3,7 +3,8 @@ _Not yet on NuGet..._
 * Plot: Added `ShowLegend()` overload that does not override the current `Orientation` (#3450) @aespitia
 * Grid: The standard grid can be accessed via `Plot.Grid` instead of `GetDefaultGrid()`
 * Grid: Allow axis-specific grid line customization (#3291, #3293) @bjschwarz, @PaxITIS
-* Legend: `Plot.Style.ColorLegend()` is deprecated. Access `Plot.Legend` properties directly as seen in the cookbook.
+* Style: `Plot.Style.ColorLegend()` is deprecated. Access `Plot.Legend` properties directly as seen in the cookbook.
+* Style: `Plot.Style.ColorAxes()` is deprecated. Call `Plot.Axes.Color()` instead.
 
 ## ScottPlot 5.0.22
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-03-16_